Refactoring verifyEmployeeId functionality
