+++ package.json
@@ -53,7 +53,7 @@
     "security-scan": "git secrets --scan && git log -p | scanrepo",
     "test": "npm run compile && jasmine",
     "comp-test": "cucumber-js component-tests/ --require build/component-tests/**/*.js",
-    "start": "npm run compile && sls offline start"
+    "start": "npm run compile && sls dynamodb install && sls offline start"
   },
   "husky": {
     "hooks": {
