+++ serverless.yml
@@ -7,6 +7,7 @@ provider:
     AZURE_AD_TENANT_ID: 6c448d90-4ca1-4caf-ab59-0a2aa67d7801
     AZURE_AD_CLIENT_ID: 09fdd68c-4f2f-45c2-be55-dd98104d4f74
     USERS_DDB_TABLE_NAME: users
+    AWS_REGION: eu-west-1
 
 package:
   individually: true
